+++ packages/web3-provider-adapter/src/get-common.ts
@@ -1,12 +1,8 @@
-import CommonOriginal, { Hardfork } from "@ethereumjs/common";
-import type { default as CommonType } from "@ethereumjs/common";
+import { Hardfork, Common } from "@ethereumjs/common";
 import { BN } from "ethereumjs-util";
-import { interopImportCJSDefault } from "node-cjs-interop";
 import { query } from "./query.js";
 
-const Common = interopImportCJSDefault(CommonOriginal);
-
-export async function getCommon(rpcUrl: string): Promise<CommonType> {
+export async function getCommon(rpcUrl: string): Promise<Common> {
   const chainId = await query<string>(rpcUrl, "eth_chainId", []);
 
   if (!chainId) {
@@ -14,7 +10,7 @@ export async function getCommon(rpcUrl: string): Promise<CommonType> {
   }
 
   return Common.custom({
-    chainId: new BN(chainId.slice(2), "hex"),
+    chainId: new BN(chainId.slice(2), "hex").toNumber(),
     defaultHardfork: Hardfork.London,
   });
 }