+++ test/e2e/legacy/web3-sign.spec.ts
@@ -9,7 +9,8 @@ const { region, keyId, rpcUrl } = getConfig();
 it("web3.js", async () => {
   const provider = new KmsProvider(rpcUrl, { region, keyIds: [keyId] });
 
-  const web3 = new Web3(provider);
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const web3 = new Web3(provider as any);
 
   const accounts = await web3.eth.getAccounts();