+++ packages/web3-provider-adapter/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/odanado/aws-kms-provider/compare/@cloud-cryptographic-wallet/web3-provider-adapter@0.1.2...@cloud-cryptographic-wallet/web3-provider-adapter@0.1.3) (2022-09-13)
+
+### Bug Fixes
+
+- **web3-provider-adapter:** fix type error ([#683](https://github.com/odanado/aws-kms-provider/issues/683)) ([c1916ee](https://github.com/odanado/aws-kms-provider/commit/c1916eec7112d3d6a2ead61f744a70afe72b0316))
+- **web3-provider-adapter:** upgrade @ethereumjs/common and @ethereumjs/tx ([#672](https://github.com/odanado/aws-kms-provider/issues/672)) ([1c85af6](https://github.com/odanado/aws-kms-provider/commit/1c85af623ae3e82cbc41107e27c6a513061b5b26))
+
 ## [0.1.2](https://github.com/odanado/aws-kms-provider/compare/@cloud-cryptographic-wallet/web3-provider-adapter@0.1.1...@cloud-cryptographic-wallet/web3-provider-adapter@0.1.2) (2022-08-18)
 
 **Note:** Version bump only for package @cloud-cryptographic-wallet/web3-provider-adapter